+++ App/Api/Client/DeLijnClientStops.cs
@@ -7,14 +7,14 @@ namespace DeLijn.Net.App.Api.Client;
 
 public sealed partial class DeLijnClient : BaseClient
 {
-    public async Task<IReadOnlyList<Stop>> GetStopsByEntityAsync(int entityId, CancellationToken? cancellationToken = null)
+    public async Task<IReadOnlyList<Stop>> GetStopsByEntityAsync(short entityId, CancellationToken? cancellationToken = null)
     {
         var responseBody = await GetAsync<StopsResponse>(ApiEndpoints.GetStopsByEntity(entityId), cancellationToken);
 
         return responseBody.Stops;
     }
 
-    public async Task<IReadOnlyList<Stop>> GetStopsByMunicipalityAsync(int municipalityId, CancellationToken? cancellationToken = null)
+    public async Task<IReadOnlyList<Stop>> GetStopsByMunicipalityAsync(short municipalityId, CancellationToken? cancellationToken = null)
     {
         var responseBody = await GetAsync<StopsResponse>(ApiEndpoints.GetStopsByMunicipality(municipalityId), cancellationToken);
 
@@ -29,7 +29,7 @@ public sealed partial class DeLijnClient : BaseClient
         return responseBody.Stops;
     }
 
-    public async Task<Stop> GetStopAsync(int entityId, int stopId, DateTimeOffset? validOn, CancellationToken? cancellationToken = null)
+    public async Task<Stop> GetStopAsync(short entityId, short stopId, DateTimeOffset? validOn, CancellationToken? cancellationToken = null)
     {
         var responseBody = await GetAsync<Stop>(ApiEndpoints.GetStop(entityId, stopId, validOn), cancellationToken);
