+++ Test/Endpoints/EndpointTests.cs
@@ -15,7 +15,7 @@ public class EndpointTests
     {
         var path = "../../../../secrets";
         var config = DeLijnConfig.FromStream(File.OpenRead(path));
-        
+
         Assert.IsNotNull(config, $"Config file {path} could not be loaded");
 
         _deLijnClient = new(config);
@@ -25,7 +25,7 @@ public class EndpointTests
     public async Task DeLijnClient_GetCoreEndpointsEndpoint_Test()
     {
         IReadOnlyList<Link> endpoints = [];
-        
+
         try
         {
 #pragma warning disable CS0618 // library coverage
@@ -42,7 +42,7 @@ public class EndpointTests
     public async Task DeLijnClient_GetAllEntitiesEndpoint_Test()
     {
         IReadOnlyList<Entity>? entities = [];
-        
+
         try
         {
             entities = await _deLijnClient.GetAllEntitiesAsync();
@@ -57,7 +57,7 @@ public class EndpointTests
     public async Task DeLijnClient_GetDiversionsEndpoint_Test()
     {
         (IReadOnlyList<Diversion> diversions, IReadOnlyList<Diversion> faults)? diversions = null;
-        
+
         try
         {
             diversions = await _deLijnClient.GetDiversionsAsync();
@@ -72,7 +72,7 @@ public class EndpointTests
     [DataRow(4, 407043)] // Valid stop
     [DataRow(100, 407043)] // Invalid entity
     [DataRow(1, 123456789)] // Invalid stop
-    public async Task DeLijnClient_GetDiversionsByStopEndpoint_Test(int entityId, int stopId)
+    public async Task DeLijnClient_GetDiversionsByStopEndpoint_Test(short entityId, short stopId)
     {
         (IReadOnlyList<Diversion> diversions, IReadOnlyList<Diversion> faults) diversions;
